updated code

+++ src/xml_bindings/init_FileOutput.cpp
@@ -16,7 +16,7 @@ namespace xml_bindings {
 void init_FileOutput(xml::Node* node, mdls::FileOutput* fileOutput)
 {
    std::string filename = get(node, "FileOutput/Path", "");
-   fileOutput->rate = xml::getDouble(node, "FileOutput/Rate", 1.0);
+   fileOutput->rate = static_cast<int>(xml::getDouble(node, "FileOutput/Rate", 1.0));
    std::cout << "Saving output to: " << filename << std::endl;
    fileOutput->fout.open(filename.c_str());
    fileOutput->frameCounter = 0;
